refactor: rearranged file directory
