+++ components/Table/Students/studentList.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
-import { Table } from "./Table/Table";
-import useStudents from "../../hooks/useStudents";
-import useAuth from "../../hooks/useAuth";
-import { STUDENT_COLUMNS } from "../Students/Table/studentColumns";
+import { Table } from "../Table";
+import useStudents from "../../../hooks/useStudents";
+import useAuth from "../../../hooks/useAuth";
+import { STUDENT_COLUMNS } from "./studentColumns";
 
 const StudentList = () => {
     const { verifyLoginData } = useAuth();
